sizeof(long)==8 Bum.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2444 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/qclib/progsint.h
@@ -103,8 +103,8 @@ void    StripExtension (char *path);
 void SetEndian(void);
 extern short   (*PRBigShort) (short l);
 extern short   (*PRLittleShort) (short l);
-extern long     (*PRBigLong) (long l);
-extern long     (*PRLittleLong) (long l);
+extern int     (*PRBigLong) (int l);
+extern int     (*PRLittleLong) (int l);
 extern float   (*PRBigFloat) (float l);
 extern float   (*PRLittleFloat) (float l);
