+++ engine/qclib/qcc.h
@@ -38,8 +38,8 @@ void qccClearHunk(void);
 
 extern short   (*PRBigShort) (short l);
 extern short   (*PRLittleShort) (short l);
-extern long     (*PRBigLong) (long l);
-extern long     (*PRLittleLong) (long l);
+extern int     (*PRBigLong) (int l);
+extern int     (*PRLittleLong) (int l);
 extern float   (*PRBigFloat) (float l);
 extern float   (*PRLittleFloat) (float l);
